pipelines/requirements.txt
pandas
minio
psycopg2-binary
python-dotenv
requests
boto3
pyarrow
fastparquet